feat: add aipreflect.ResourceDescriptor
A parsed descriptor for a resource.

Moves resourcetype.Name into the aipreflect package to enable the Type
field to be of the type aipreflect.ResourceTypeDescriptor.

Some fields can not be set with the information in the protobuf
descriptor (such as parent information, file and message information,
etc).

This information will be set by a resource registry where resources can
be added and their provenance tracked.