fix(filtering): set expr.ID on nested AND expressions
Previously nested AND expressions (ex: "a AND b AND c") would not set an
ID on the innermost AND expression. This commit changes the "folding" of
a sequence of AND expressions to be similar to OR expressions.

Also adds a test to ensure that all tested expressions produce
subexpressions with unique IDs.

This fixes an issue where type checking fails because of duplicate IDs.