fix(fieldmask): dont populate message type when updating from nil
Previously when calling `Update` with paths in the fieldmask which
covered message fields, this method would allocate a new empty message
instead of nil.

Concretely this meant that timestamps could not be properly unset with a
fieldmask, as they would not become nil but rather `{ Seconds: 0, Nanos:
0}` which is a valid timestamp.

This also changes how oneofs are Updated when the source does not
specify a kind. This is the correct behavior.