feat(filtering/exprs): package for working with exprs
When using filters in APIs it is typical to validate the filter (for
example resource names) and transform it to a database schema.

The `filtering` package already exposes functionality for walking
(`filtering.Walk`) and transforming (`filtering.ApplyMacros`) exprs.
When using either of those methods some functions should be applied to
only parts of the expression tree, which typically results multiple if
statements parsing the expression.

The `exprs` package exposes several `Matcher` functions which can be
used to match parts of the expression tree. See the Go example for an
example use case.