feat: add support for generating service name specific resource names
When 2 resource names contain the same type but different service names,
the code generated becomes invalid since we do not take into account the
service name.
This adds support for including the service name in the generation and
also to remove a suffix from the service name before generation for
slightly more user friendly struct names in systems where the most of
the service name is the same and only parts change (for example
`book.library.com` and `publisher.library.com`).

Test data slightly restructured to allow for testing the flags