fix: avoid asserting on exact page token value
The page token value changes as of Go 1.21, likely due to changes in Gob
encoding. Old values should still be unmarshaled properly by Go 1.21, so
shouldn't be an issue.

Avoid asserting on the exact page token value in a test, and instead log
if it is empty or not, as the test will produce different outputs
depending on the Go version used.