feat: introduce experimental improved filtering
While aip-go has fairly robust ways of filtering as per
[AIP-160](https://google.aip.dev/160), it is somewhat cumbersome to use.

When working with AIP one is generally working with data models defined
in protobuf files so it makes sense that filter expressions are written
towards the definition of these models however ultimately the filter is
likely translated to some other system for actual data retrieval (like
a SQL database).

In order to do this, aip-go has supported parsing of filters (through
declarations) and macros which allow transforming the parsed filter
expression into an output expression which is more suitable to whatever
underlying system is being used.

Using current functionality, this might look like this:

1. Declare what filters are supported and their types (from protobuf
   fields)
2. Declare macros which transform the parsed filter
3. Declare what filters are supported in the underlying system (this is
   needed because macros only work with typeless expressions so there
   needs to be type definitions that match the transformed filter in
   order to better support, for example, transpiling to SQL.


Some drawbacks from doing this:

1. Writing and declaring fields and their types by hand is prone to
   error and feel like a chore. This information is already present in
   the proto files anyway.
2. Once one has declared what filters are supported (step 1) and
   declared what transformations are wanted (step 2), having to write the
   validation declarations (step 3) feels like a bit of a redundant step,
   specially when a database table may have columns named exactly like
   the proto fields.

This commit attempts to improve upon this workflow while avoiding any
breaking changes to the API.

More specifically:

Introduce a new `Cursor.ReplaceWithDeclarations` method on the Cursor type
which can be used inside a macro. This allows a macro to not only
replace a matching expression but also define any new types introduced
by the replacement expression.

Pass the original filter declarations in the Cursor type along with a
new `Cursor.LookupIdentType` method which allows a macro to lookup
existing identifier types (specifically from the original declarations
and not from other macros). This can be used, for example, to replace an
identifier name with another while keeping the existing type the same.



Introduce a new `Filter.ApplyMacros` method in the Filter type. This is done
to work around the fact that it was not possible to change the
existing `ApplyMacros` without breaking existing behaviour.
`filter.ApplyMacros` works slightly different in that it does not expect
declarations to validate against and instead validates against both the
original declarations and any declarations introduced with
`ReplaceWithDeclarations`.
Note that this means that any declarations passed in `ParseFilter` that
are not replaced/modified are considered valid in the resulting filter
from `Filter.ApplyMacros`.

Introduce a new `ProtoDeclarations` function which can take any
`proto.Message` and a list of options to automatically derive filter
declarations. The only available option for now is
`WithFilterableFields` in which one can pass a list of field names that
should be included in the final declarations (types are derived directly
from the proto message).

NOTE: All of this is very much experimental and marked as such while we
iterate through the API and any eventual bugs